create dummy secret.json for docker if it doesnt exist

+++ frontend/docker_entrypoint.sh
@@ -10,8 +10,20 @@ npm install
 
 PATH=$(npm bin):$PATH bower install
 
-cp config/envs/docker.json config/envs/devel.json
-cp config/envs/docker.json config/envs/production.json
+if [ ! -f config/envs/devel.json ]; then
+    echo 'create config/envs/devel.json.'
+  cp config/envs/docker.json config/envs/devel.json
+fi
+
+if [ ! -f config/envs/production.json ]; then
+    echo 'create config/envs/production.json.'
+    cp config/envs/docker.json config/envs/production.json
+fi
+
+if [ ! -f secret.json ]; then
+    echo 'create secret.json.'
+    cp secret.json.example secret.json
+fi
 
 PATH=$(npm bin):$PATH grunt serve
